INS-108: split RPC types into a package for client and server to use
