+++ logicrunner/goplugin/girpc/girpc.go
@@ -1,0 +1,17 @@
+package girpc
+
+import "github.com/insolar/insolar/logicrunner"
+
+// CallReq is a set of arguments for Call RPC in the runner
+type CallReq struct {
+	Object logicrunner.Object
+	Method string
+	Args   logicrunner.Arguments
+}
+
+// CallResp is response from Call RPC in the runner
+type CallResp struct {
+	Data []byte
+	Ret  logicrunner.Arguments
+	Err  error
+}
